Wrap debug code in process.env checks in the key module
Part of https://github.com/canjs/canjs/issues/4011

+++ key/key.js
@@ -9,9 +9,11 @@ module.exports = function keyObservable(root, keyPath) {
 
 	// Some variables used to build the dependency/mutation graph
 	//!steal-remove-start
-	var lastIndex = keyPathParts.length - 1;
-	var lastKey;// This stores the last part of the keyPath, e.g. “key” in “outer.inner.key”
-	var lastParent;// This stores the object that the last key is on, e.g. “outer.inner” in outer: {inner: {"key": "value"}}
+	if (process.env.NODE_ENV !== 'production') {
+		var lastIndex = keyPathParts.length - 1;
+		var lastKey;// This stores the last part of the keyPath, e.g. “key” in “outer.inner.key”
+		var lastParent;// This stores the object that the last key is on, e.g. “outer.inner” in outer: {inner: {"key": "value"}}
+	}
 	//!steal-remove-end
 
 	var observation = new Observation(function() {
@@ -21,13 +23,15 @@ module.exports = function keyObservable(root, keyPath) {
 		canKey.walk(root, keyPathParts, function(keyData, i) {
 			if (i === lastIndex) {
 				//!steal-remove-start
-				// observation is mutating keyData.parent
-				if (lastParent && (keyData.key !== lastKey || keyData.parent !== lastParent)) {
-					canReflectDependencies.deleteMutatedBy(lastParent, lastKey, observation);
+				if (process.env.NODE_ENV !== 'production') {
+					// observation is mutating keyData.parent
+					if (lastParent && (keyData.key !== lastKey || keyData.parent !== lastParent)) {
+						canReflectDependencies.deleteMutatedBy(lastParent, lastKey, observation);
+					}
+					lastKey = keyData.key;
+					lastParent = keyData.parent;
+					canReflectDependencies.addMutatedBy(lastParent, lastKey, observation);
 				}
-				lastKey = keyData.key;
-				lastParent = keyData.parent;
-				canReflectDependencies.addMutatedBy(lastParent, lastKey, observation);
 				//!steal-remove-end
 
 				value = keyData.value;
@@ -37,17 +41,23 @@ module.exports = function keyObservable(root, keyPath) {
 		return value;
 	});
 
-	return canReflect.assignSymbols(observation, {
-		//!steal-remove-start
-		"can.getName": function getName() {
+	var symbolsToAssign = {
+		"can.setValue": function(newVal) {
+			canKey.set(root, keyPathParts, newVal);
+		}
+	};
+
+	//!steal-remove-start
+	if (process.env.NODE_ENV !== 'production') {
+
+		// Debug name
+		symbolsToAssign["can.getName"] = function getName() {
 			var objectName = canReflect.getName(root);
 			return "keyObservable<" + objectName + "." + keyPath + ">";
-		},
-		//!steal-remove-end
+		};
 
 		// Register what this observable changes
-		//!steal-remove-start
-		"can.getWhatIChange": function getWhatIChange() {
+		symbolsToAssign["can.getWhatIChange"] = function getWhatIChange() {
 			return {
 				mutate: {
 					keyDependencies: new Map([
@@ -55,11 +65,9 @@ module.exports = function keyObservable(root, keyPath) {
 					])
 				}
 			};
-		},
-		//!steal-remove-end
+		};
+	}
+	//!steal-remove-end
 
-		"can.setValue": function(newVal) {
-			canKey.set(root, keyPathParts, newVal);
-		}
-	});
+	return canReflect.assignSymbols(observation, symbolsToAssign);
 }
